+++ asdf_astropy/extensions.py
@@ -35,6 +35,7 @@ from .converters.unit.equivalency import EquivalencyConverter
 from .converters.unit.magunit import MagUnitConverter
 from .converters.unit.quantity import QuantityConverter
 from .converters.unit.unit import UnitConverter
+from .converters.wcs.highlevelwrapper import HighLevelWCSWrapperConverter
 from .converters.wcs.slicedwcs import SlicedWCSConverter
 from .converters.wcs.wcs import WCSConverter
 
@@ -530,6 +531,7 @@ ASTROPY_CONVERTERS = [
     UncertaintyConverter(),
     WCSConverter(),
     SlicedWCSConverter(),
+    HighLevelWCSWrapperConverter(),
 ]