+++ CHANGES.rst
@@ -9,6 +9,7 @@
 
 - drop support for numpy 1.25. [#285]
 - require asdf 3.3 and drop support for astropy 5.2 [#297]
+- Add support for ``astropy.wcs.wcsapi.HighLevelWCSWrapper``. [#296]
 
 0.8.0 (2025-06-11)
 ------------------